+++ lib/biotcm/modules.rb
@@ -1,0 +1,7 @@
+# encoding: UTF-8
+
+# Namespace of all modules
+module BioTCM::Modules
+  autoload(:Utility, 'biotcm/modules/utility')
+  autoload(:WorkingDir, 'biotcm/modules/workingdir')
+end