+++ HISTORY.md
@@ -1,6 +1,27 @@
 History
 ===============
 
+### By now 2014-03-20
+* 4 minor enhancement
+	* Improved BioTCM::Table row insertion
+	* Improved BioTCM::Databases::Cipher performance
+	* Improved our benchmark suite
+	* Added two setter methods for BioTCM::Table
+
+### 0.0.2 2014-03-16
+* 3 major enhancement
+	* Added BioTCM::Logger
+	* Added BioTCM::Databases::HGNC
+	* Added BioTCM::Databases::Cipher
+
+### 0.0.1 2014-03-14
+* 2 major enhancement
+	* Completed the design of architectures
+	* Completed following modules and classes
+		* Added BioTCM::Modules::Utility
+		* Added BioTCM::Modules::WorkingDir
+		* Added BioTCM::Table
+
 ### 0.0.0 2014-03-06
 * 1 major enhancement
 	* Built biotcm gem